feat: split container component
